use timeout command if it exists

+++ timeout.sh
@@ -1,3 +1,9 @@
 #!/bin/bash
 
-( cmdpid=$BASHPID; (sleep $1; kill $cmdpid) & exec "${@:2}" )
+timeout=`which timeout`
+if [ -f $timeout ]; then
+    $timeout $@
+else
+    ( cmdpid=$BASHPID; (sleep $1; kill $cmdpid) & exec "${@:2}" )
+    exit 0
+fi
